+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -236,6 +236,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             timePickerElementRadius: controller.timePickerElementRadius,
             dialogElevation: controller.dialogElevation,
             // SnackBar settings.
+            snackBarRadius: controller.snackBarBorderRadius,
+            snackBarElevation: controller.snackBarElevation,
             snackBarBackgroundSchemeColor: controller.snackBarSchemeColor,
             snackBarActionSchemeColor: controller.snackBarActionSchemeColor,
             // AppBar settings.